requirements.txt
flask
flask-cors
requests
xmltodict
python-dotenv
slack-bolt
